+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 * Allow the user to switch between vector and satellite image map.
 * Remove `InfoBlock` mechanism for marker details overlay. (popup or teaser)
+  * Therefore marker details are - at the moment - not supported in the Web Component.
 
 ## [2.6.0](https://code.sbb.ch/projects/KI_ROKAS/repos/journey-maps-client/compare/commits?targetBranch=refs%2Ftags%2F2.5.0&sourceBranch=refs%2Ftags%2F2.6.0&targetRepoId=46287) - 2021-11-29